Move code from original places to:     utils/unicode_converter.cpp     utils/conversion_context.cpp     utils/utils.h
