+++ driver/CMakeLists.txt
@@ -54,6 +54,8 @@ unset (WORKAROUND_ENABLE_NO_UNICODE_CHARS_IN_COLUMN_NAMES_IN_TESTS)
 add_library (${libname}-impl STATIC
     utils/type_parser.cpp
     utils/type_info.cpp
+    utils/unicode_converter.cpp
+    utils/conversion_context.cpp
 
     config/config.cpp
